reddit-worker/requirements.txt
praw==7.8.1
supabase==2.18.1
schedule==1.2.2
python-dotenv==1.1.1